backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic>=2.10.3
python-dotenv==1.0.0
selenium==4.15.2
openai==1.3.7
mistralai==1.8.2
python-multipart==0.0.6
aiofiles==23.2.1
numpy>=1.26.0
pytest==7.4.3
pytest-asyncio==0.21.1
flake8==6.1.0
beautifulsoup4==4.12.2
requests==2.31.0
httpx==0.28.1
python-dateutil>=2.8.2
typing-inspection>=0.4.0
eval-type-backport>=0.2.0